requirements.txt
asgiref==3.4.1
captcha==0.3
certifi==2021.10.8
charset-normalizer==2.0.12
dataclasses==0.8
distlib==0.3.6
Django==3.2.15
django-ckeditor==6.3.2
django-js-asset==2.0.0
django-mptt==0.13.4
docopt==0.6.2
filelock==3.4.1
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
multicolorcaptcha==1.2.0
Pillow==8.4.0
pipreqs==0.4.11
platformdirs==2.4.0
pyTelegramBotAPI==4.4.0
pyTelegramBotCAPTCHA==1.1.4
pytz==2022.2.1
requests==2.27.1
sqlparse==0.4.2
telebot==0.0.4
typing_extensions==4.1.1
urllib3==1.26.8
virtualenv==20.16.3
yarg==0.1.9
zipp==3.6.0
